moved machine names/aliases, logbook credentials, and server addresses from fdp_globals.py to datasources.py
